+++ library-no-op/src/main/kotlin/com/chuckerteam/chucker/api/ChuckerCollector.kt
@@ -16,5 +16,6 @@ public class ChuckerCollector @JvmOverloads constructor(
     public fun writeTransactions(
         context: Context,
         startTimestamp: Long?,
+        exportFormat: ExportFormat = ExportFormat.LOG
     ): Uri? = null
 }