+++ library-no-op/src/main/kotlin/com/chuckerteam/chucker/api/ExportFormat.kt
@@ -1,0 +1,12 @@
+package com.chuckerteam.chucker.api
+
+/**
+ * The supported export format for transactions file exports.
+ */
+public enum class ExportFormat(public val extension: String) {
+    /** LOG Format with txt extension */
+    LOG("txt"),
+
+    /** HAR format with har extension */
+    HAR("har")
+}