+++ library/src/main/kotlin/com/chuckerteam/chucker/api/ChuckerCollector.kt
@@ -2,9 +2,14 @@ package com.chuckerteam.chucker.api
 
 import android.content.Context
 import android.net.Uri
+import com.chuckerteam.chucker.R
 import com.chuckerteam.chucker.internal.data.entity.HttpTransaction
+import com.chuckerteam.chucker.internal.data.har.log.Creator
 import com.chuckerteam.chucker.internal.data.repository.RepositoryProvider
+import com.chuckerteam.chucker.internal.support.HarUtils
+import com.chuckerteam.chucker.internal.support.JsonConverter
 import com.chuckerteam.chucker.internal.support.NotificationHelper
+import com.chuckerteam.chucker.internal.support.TransactionDetailsHarSharable
 import com.chuckerteam.chucker.internal.support.TransactionListDetailsSharable
 import com.chuckerteam.chucker.internal.support.writeToFile
 import kotlinx.coroutines.Dispatchers
@@ -74,11 +79,13 @@ public class ChuckerCollector @JvmOverloads constructor(
      * @param context Application context
      * @param startTimestamp The timestamp to read transactions from. Passing null means
      * transactions will not be limited by timestamp
+     * @param exportFormat The export format: LOG or HAR
      * @return The content uri of a file with the transactions in or null if the export failed.
      */
     public fun writeTransactions(
         context: Context,
         startTimestamp: Long?,
+        exportFormat: ExportFormat = ExportFormat.LOG
     ): Uri? {
         val transactions =
             RepositoryProvider.transaction().getTransactionsInTimeRange(startTimestamp)
@@ -86,10 +93,27 @@ public class ChuckerCollector @JvmOverloads constructor(
             return null
         }
 
-        val sharableTransactions = TransactionListDetailsSharable(transactions, encodeUrls = false)
+        val sharableTransactions =  when (exportFormat) {
+            ExportFormat.LOG -> {
+                TransactionListDetailsSharable(transactions, encodeUrls = false)
+            }
+            ExportFormat.HAR -> {
+                TransactionDetailsHarSharable(
+                    JsonConverter.nonNullSerializerInstance
+                        .toJson(
+                            HarUtils.fromHttpTransactions(
+                                transactions, Creator(
+                                    context.getString(R.string.chucker_name),
+                                    context.getString(R.string.chucker_version),
+                                )
+                            )
+                        )
+                )
+            }
+        }
         return sharableTransactions.writeToFile(
             context = context,
-            fileName = "api_transactions.txt",
+            fileName = "api_transactions.${exportFormat.extension}",
         )
     }